feat: Delete Notification 개별 구현

+++ feature/main/src/main/java/com/kitching/main/view/other/tab/NoticeNotificationListScreen.kt
@@ -19,7 +19,7 @@ import com.kitching.main.view.other.item.NoticeNotificationCardItem
 @Composable
 fun NoticeNotificationListScreen(
     noticeNotifications: List<NoticeNotification>,
-    onDeleteClick: () -> Unit
+    onDeleteClick: (Long) -> Unit
 ) {
     if (noticeNotifications.isEmpty()) {
         EmptyScreen(stringResource(R.string.notice_notification_empty_text))
